Performance improvements
drawImage is faster. A bitmap is the same as GPU memory by design. Much faster than writing back pixels in GPU memory with CPU. Faster font loading and better initialization. It would be nice if we could get the canvas memory bit map to image element rather than writing it back out as a base64 string for the bit map data to create the image from a data URL. Since this is only done once or twice as the hex editor is made larger in height this is not a performance bottle neck.